+++ docs/CHANGELOG.md
@@ -30,4 +30,5 @@
 - Exposes C API using graalvm to allow the lib to communicate with other programs 
 - Fix an error for C API fetch_rows
 - Replace Poly2Tri by Tinfour library, signature of ST_ConstrainedDelaunay now have a minDistancePoint parameter for merging input coordinates
+- Fix storage of Boolean and Byte type of the Flat GeoBuffer storage driver issue #1437
 - Fix transformation from 27700 to 4326